% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptualPlotWidth.R
\name{conceptualPlotWidth}
\alias{conceptualPlotWidth}
\title{Conceptual plot of an antibody profile regarding Width.}
\usage{
conceptualPlotWidth(
  i,
  ids,
  df,
  part_col,
  weight_col,
  val_col,
  group_col = NULL,
  groups = NULL,
  age_col = NULL,
  var_trans = 1,
  base = 2,
  adj = 10,
  mode = "mean",
  adjust = NULL,
  cutoff = 3,
  x.min = 1965,
  x.max = 2015,
  x.diff.1 = 5,
  x.diff.2 = 1,
  y.min = 0,
  y.max = 8,
  y.diff = 1,
  legends = c("B"),
  xlab = "Year of isolation",
  ylab = "Titers",
  color.cutoff = "#e6a953",
  color.1 = "#e9d5bd",
  color.2 = "#b87624",
  color.line = "#999999",
  line.type.cutoff = 2,
  point.type = 19,
  cex = 1.5,
  cex.lab = 2,
  cex.main = 2,
  cex.axis = 1.5,
  mar = c(6, 6, 5, 2),
  mgp = c(3, 1, 4),
  width.dataline = 2,
  width.cutoff = 2,
  legend.font = 2
)
}
\arguments{
\item{i, }{panel i}

\item{ids, }{a vector of randomly selected participant id}

\item{df, }{dataframe}

\item{part_col}{a vector identifying participant id}

\item{weight_col}{a vector of the distance used to sorting titers. Eg. year of isolation or genetic distance}

\item{val_col}{a vector of target values. Eg. titers}

\item{group_col}{a vector identifying the sample group, default is NULL}

\item{groups}{a list of group numbers, default is NULL}

\item{age_col}{a vector of the age of participants, used for plotting the title of the plot, default is NULL}

\item{var_trans}{whether the input titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{base}{log base used to calculate the log transformed value, default is 2}

\item{adj}{dividing factor before taking log transformation, default is 10}

\item{mode}{how to calculate values if weight_col is repeated, default is "mean"}

\item{adjust}{how to adjust titers when they are too small to plot, default is NULL}

\item{cutoff}{the cutoff for the width calculation, default is 3 and is log-transformed}

\item{x.min}{the start point of the x-axis, default is 1965}

\item{x.max}{the end point of the x-axis, default is 2015}

\item{x.diff.1}{the larger interval of the segments on the x-axis, default is 5}

\item{x.diff.2}{the smaller interval of the segments on the x-axis, default is 1}

\item{y.min}{the start point of the x-axis, default is 0}

\item{y.max}{the end point of the x-axis, default if 10}

\item{y.diff}{the interval of the segments on the y-axis, default is 1}

\item{legends}{the legends of the plot, default is NULL}

\item{xlab}{the label of the x-axis, default is "Year of isolation"}

\item{ylab}{the label of the y-axis, default is "Titers"}

\item{color.cutoff}{the color of the horizontal cutoff line, default is "#e6a953"}

\item{color.1}{the color of the horizontal bar above the 0-axis where the titer values are smaller than the cutoff, default is '#e9d5bd'}

\item{color.2}{the color of the horizontal bar above the 0-axis where the titer values are larger than the cutoff, default is '#b87624'}

\item{color.line}{the color of the curve, default is "#999999"}

\item{line.type.cutoff}{the type of the cutoff line, default is 2}

\item{point.type}{the type of the data points above the cutoff line, default is 19(a filled circle)}

\item{cex}{the font size of the legend, default is 1.5}

\item{cex.lab}{the font size of the lab of x- and y-axis, default is 2.0}

\item{cex.main}{the font size of the title of the figures, default is 2.0}

\item{cex.axis}{the font size of the segments of x- and y-axis, default is 1.5}

\item{mar}{the margin of the figure, default is c(6,6,5,2)}

\item{mgp}{the layout of the figure, default is c(3,1,4)}

\item{width.dataline}{the width of the curve, default is 2}

\item{width.cutoff}{the width of the cutoff line, default is 2}

\item{legend.font}{the font size of the legend, default is 2}
}
\value{
the conceptual plot regarding AUC
}
\description{
Conceptual plot of an antibody profile regarding Width.
}
