% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransform.R
\name{dataTransform}
\alias{dataTransform}
\title{Function to transform data frame}
\usage{
dataTransform(
  df,
  panel,
  col_start = NULL,
  col_end = NULL,
  strain_col = "strain",
  val_col = "titer",
  group_cols = NULL,
  group_col = NULL,
  def = "_",
  drop = 1
)
}
\arguments{
\item{df}{a data frame}

\item{panel}{the panel of data transformation, panel = 1 means transpose some columns, panel = 2 means combining group columns}

\item{col_start}{the column number from which the transformation starts, default is NULL}

\item{col_end}{the column number at which the transformation ends, default is NULL}

\item{strain_col}{the column name of the strains in the output, default is "strain"}

\item{val_col}{the column name of the titer values in the output, default is "titer"}

\item{group_cols}{the group columns intended to be combined, default is NULL}

\item{group_col}{the new group column name, default is NULL}

\item{def}{the default linking between group column names to form the new column name, default is "_"}

\item{drop}{whether to drop the original group columns, default is TRUE}
}
\value{
a transformed data frame, with the column in the range \link{col_start, col_end} shifted to two columns
}
\description{
Function to transform data frame
}
