% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualPlot.R
\name{dualPlot}
\alias{dualPlot}
\title{Function to plot comparisons between two antibody profiles.}
\usage{
dualPlot(
  i,
  ids.1,
  ids.2,
  df,
  part_col,
  weight_col,
  val_col,
  group_col = NULL,
  groups.1 = NULL,
  groups.2 = NULL,
  age_col = NULL,
  birth_col = NULL,
  sample_years.1 = NULL,
  sample_years.2 = NULL,
  mode = "mean",
  var_trans = 1,
  base = 2,
  adj = 10,
  adjust = NULL,
  x.min = 1965,
  x.max = 2015,
  x.diff.1 = 5,
  x.diff.2 = 1,
  y.min = 0,
  y.max = 10,
  y.diff = 1,
  legends = c("A"),
  xlab = "Distance",
  ylab = "Titers",
  color.1 = "#3f93aa",
  color.2 = "#fa8072",
  color.main = "#867b7b",
  color.change.1 = "#590059",
  color.change.2 = "#8BAD23",
  line.type.1 = 1,
  line.type.2 = 3,
  point.type.1 = 1,
  point.type.2 = 2,
  line.type.birth.1 = 2,
  line.type.birth.2 = 4,
  cex = 1.5,
  cex.lab = 2,
  cex.main = 2,
  cex.axis = 1.5,
  mar = c(6, 6, 5, 2),
  mgp = c(3, 1, 4),
  width.dataline = 2,
  width.points = 1.5,
  legend.font = 2
)
}
\arguments{
\item{i, }{panel i}

\item{df, }{dataframe}

\item{part_col}{a vector identifying participant id, default is "id"}

\item{weight_col}{a vector of the distance used to sorting titers. Eg. year of isolation or genetic distance}

\item{val_col}{a vector of target values. Eg. titers}

\item{group_col}{a vector identifying the sample group, default is NULL}

\item{groups.1}{a list of group numbers of the first group of profiles, default is NULL}

\item{groups.2}{a list of group numbers of the second group of profiles, default is NULL}

\item{age_col}{a vector of the age of participants, used for plotting the title of the plot, default is NULL}

\item{birth_col}{a vector identifying the birth year of the participants, used for plotting the birth year of the participant, default is NULL}

\item{sample_years.1}{a list of sample years for plotting vertical bars of the first profiles, default is NULL}

\item{sample_years.2}{a list of sample years for plotting vertical bars of the second profiles, default is NULL}

\item{mode}{how to calculate values if weight_col is repeated, default is "mean"}

\item{var_trans}{whether the input titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{base}{log base used to calculate the log transformed value, default is 2}

\item{adj}{dividing factor before taking log transformation, default is 10}

\item{adjust}{how to adjust titers when they are too small to plot, default is NULL}

\item{x.min}{the start point of the x-axis, default is 1965}

\item{x.max}{the end point of the x-axis, default is 2015}

\item{x.diff.1}{the larger interval of the segments on the x-axis, default is 5}

\item{x.diff.2}{the smaller interval of the segments on the x-axis, default is 1}

\item{y.min}{the start point of the x-axis, default is 0}

\item{y.max}{the end point of the x-axis, default if 10}

\item{y.diff}{the interval of the segments on the y-axis, default is 1}

\item{legends}{the legends of the plot, default is "A"}

\item{xlab}{the label of the x-axis, default is "Distance"}

\item{ylab}{the label of the y-axis, default is "Titers"}

\item{color.1}{the color of the curve, the data points, the vertical bar of sample year, and the vertical line of birth year, for the first group. default is "#3f93aa"}

\item{color.2}{the color of the curve, the data points, the vertical bar of sample year, and the vertical line of birth year, for the second group. default is "#fa8072"}

\item{color.main}{the color of the area under the curves, default is "#867b7b"}

\item{color.change.1}{the color of the area above the curve of the first group and below the curve of the second group, default is "#590059"}

\item{color.change.2}{the color of the area above the curve of the second group and below the curve of the first group, default is "#8BAD23"}

\item{line.type.1}{the line type of the curve of the first group, default is 1}

\item{line.type.2}{the line type of the curve of the second group, default is 3}

\item{point.type.1}{the point type of the data points of the first group, default is 1}

\item{point.type.2}{the point type of the data points of the second group, default is 2}

\item{line.type.birth.1}{the line type of the birth line of the first group, default is 2}

\item{line.type.birth.2}{the line type of the birth line of the second group, default is 4}

\item{cex}{the font size of the legend, default is 1.5}

\item{cex.lab}{the font size of the lab of x- and y-axis, default is 2.0}

\item{cex.main}{the font size of the title of the figures, default is 2.0}

\item{cex.axis}{the font size of the segments of x- and y-axis, default is 1.5}

\item{mar}{the margin of the figure, default is c(6,6,5,2)}

\item{mgp}{the layout of the figure, default is c(3,1,4)}

\item{width.dataline}{the width of the curve, default is 2.0}

\item{width.points}{the width of the data points, default is 1.5}

\item{legend.font}{the font size of the legend, default is 2}

\item{ids, }{a vector of randomly selected participant id}
}
\value{
a plot of individual antibody profile
}
\description{
Function to plot comparisons between two antibody profiles.
}
