% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.R
\name{gini}
\alias{gini}
\title{Calculate the gini coefficient of the titers}
\usage{
gini(titer, input.log.trans, base = 2, adj = 10, adjust = NULL)
}
\arguments{
\item{titer}{a vector of titer values}

\item{input.log.trans}{whether the input titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{base}{log base used to calculate the log transformed value, default is 2}

\item{adj}{dividing factor before taking log transformation, default is 10}

\item{adjust}{how to adjust titers, default is the minimum of titers - 1}
}
\value{
gini
}
\description{
Calculate the gini coefficient of the titers
}
