% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini_data.R
\name{gini_data}
\alias{gini_data}
\title{Function to format data and calculate gini}
\usage{
gini_data(
  data,
  part_col,
  val_col,
  group_col = NULL,
  var_trans = TRUE,
  base = 2,
  adj = 10,
  adjust = NULL,
  gini_col = "gini_coefficient"
)
}
\arguments{
\item{data}{a data frame containing the necessary columns (participant_id, titer)}

\item{part_col}{a vector identifying participant id}

\item{val_col}{a vector of target values. Eg. titers}

\item{group_col}{a vector identifying the sample group, default is NULL}

\item{var_trans}{whether the input titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{base}{log base used to calculate the log transformed value, default is 2}

\item{adj}{dividing factor before taking log transformation, default is 10}

\item{adjust}{how to adjust titers, default is the minimum of titers - 1}

\item{gini_col}{the column name of the gini coefficient column, default is "gini_coefficient"}
}
\value{
a data frame containing the individual gini values
}
\description{
Function to format data and calculate gini
}
\examples{
\donttest{
data(Fonville)
gini_data(Fonville, part_col = "Subject Number", val_col = "titer",
         group_col = "Sample Year")
}

}
