% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giniData.R
\name{giniCal}
\alias{giniCal}
\title{Calculate the gini for a given participant}
\usage{
giniCal(
  i,
  j,
  data,
  part_col,
  val_col,
  group_col,
  var_trans,
  base,
  adj,
  adjust = NULL
)
}
\arguments{
\item{i}{a participant ID}

\item{j}{a sample group number}

\item{data}{the input data frame}

\item{part_col}{a vector identifying participant id}

\item{val_col}{a vector of target values. Eg. titers}

\item{group_col}{a vector identifying the sample group}

\item{var_trans}{whether the input titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{base}{a parameter passed to the \code{gini} function}

\item{adj}{a parameter passed to the \code{gini} function}

\item{adjust}{how to adjust titers, default is the minimum of titers - 1}

\item{weight_col}{a vector of the distance used to sorting titers. Eg. year of isolation or genetic distance}
}
\value{
the gini value for the given participant
}
\description{
Calculate the gini for a given participant
}
