% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giniData.R
\name{giniData}
\alias{giniData}
\title{Function to format data and calculate gini}
\usage{
giniData(
  data,
  part_col,
  val_col,
  group_col = NULL,
  var_trans = 1,
  base = 2,
  adj = 10,
  adjust = NULL,
  gini_col = "gini_coefficient"
)
}
\arguments{
\item{data}{a data frame containing the necessary columns (participant_id, titer)}

\item{part_col}{a vector identifying participant id}

\item{val_col}{a vector of target values. Eg. titers}

\item{group_col}{a vector identifying the sample group, default is NULL}

\item{var_trans}{whether the input titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{base}{log base used to calculate the log transformed value, default is 2}

\item{adj}{dividing factor before taking log transformation, default is 10}

\item{adjust}{how to adjust titers, default is the minimum of titers - 1}

\item{gini_col}{the column name of the gini coefficient column, default is "gini_coefficient"}

\item{weight_col}{a vector of the distance used to sorting titers. Eg. year of isolation or genetic distance}
}
\value{
a data frame containing the individual gini values
}
\description{
Function to format data and calculate gini
}
