% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt.R
\name{gmt}
\alias{gmt}
\title{Calculate geometric mean titer against a panel of viruses}
\usage{
gmt(titer, input.log.trans = TRUE, base = 2, adj = 10, output.log.trans = TRUE)
}
\arguments{
\item{titer}{titer value}

\item{input.log.trans}{whether the input titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{base}{log base used to calculate the log transformed value, default is 2}

\item{adj}{dividing factor before taking log transformation, default is 10}

\item{output.log.trans}{whether the output titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}
}
\value{
GMT
}
\description{
Calculate geometric mean titer against a panel of viruses
}
