% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmtData.R
\name{gmtCal}
\alias{gmtCal}
\title{Calculate the GMT for a given participant}
\usage{
gmtCal(
  i,
  j,
  data,
  part_col,
  val_col,
  group_col,
  var_trans,
  output_trans,
  base,
  adj
)
}
\arguments{
\item{i}{a participant ID}

\item{j}{a sample group number}

\item{data}{the input data frame}

\item{part_col}{a vector identifying participant id}

\item{val_col}{a vector of target values. Eg. titers}

\item{group_col}{a vector identifying the sample group}

\item{var_trans}{whether the input titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{output_trans}{whether the output titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{base}{a parameter passed to the `gmt` function}

\item{adj}{a parameter passed to the `gmt` function}
}
\value{
the GMT value for the given participant
}
\description{
Calculate the GMT for a given participant
}
