% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmtData.R
\name{gmtData}
\alias{gmtData}
\title{Function to format data and calculate GMT}
\usage{
gmtData(
  data,
  part_col,
  val_col,
  group_col = NULL,
  var_trans = 1,
  output_trans = 1,
  base = 2,
  adj = 10
)
}
\arguments{
\item{data}{a data frame containing the necessary columns (participant_id, titer)}

\item{part_col}{a vector identifying participant id}

\item{val_col}{a vector of target values. Eg. titers}

\item{group_col}{a vector identifying the sample group, default is NULL}

\item{var_trans}{whether the input titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{output_trans}{whether the output titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{base}{log base used to calculate the log transformed value, default is 2}

\item{adj}{dividing factor before taking log transformation, default is 10}
}
\value{
a data frame containing the individual GMT values
}
\description{
Function to format data and calculate GMT
}
