% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtosis_data.R
\name{kurtosis_data}
\alias{kurtosis_data}
\title{Function to format data and calculate kurtosis}
\usage{
kurtosis_data(
  data,
  part_col,
  weight_col,
  val_col,
  group_col = NULL,
  var_trans = TRUE,
  base = 2,
  adj = 10,
  mode = NULL,
  adjust = NULL,
  kurt_col = "Kurtosis"
)
}
\arguments{
\item{data}{a data frame containing the necessary columns (participant_id, titer)}

\item{part_col}{a vector identifying participant id}

\item{weight_col}{a vector of the distance used to sorting titers. Eg. year of isolation or genetic distance}

\item{val_col}{a vector of target values. Eg. titers}

\item{group_col}{a vector identifying the sample group, default is NULL}

\item{var_trans}{whether the input titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{base}{log base used to calculate the log transformed value, default is 2}

\item{adj}{dividing factor before taking log transformation, default is 10}

\item{mode}{how to calculate values if weight_col is repeated, default is NULL}

\item{adjust}{how to adjust titers, default is the minimum of titers}

\item{kurt_col}{the column name of the Kurtosis column, default is "Kurtosis"}
}
\value{
a data frame containing the individual kurtosis values
}
\description{
Function to format data and calculate kurtosis
}
\examples{
\donttest{
data(Fonville)
kurtosis_data(Fonville, part_col = "Subject Number", weight_col = "isolation_year",
             val_col = "titer", group_col = "Sample Year")
}

}
