% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_titer_data.R
\name{max_titer_data}
\alias{max_titer_data}
\title{Function to format data and calculate max_titer}
\usage{
max_titer_data(
  data,
  part_col,
  val_col,
  group_col = NULL,
  var_trans = TRUE,
  output_trans = TRUE,
  base = 2,
  adj = 10,
  max_titer_col = "Max_titer"
)
}
\arguments{
\item{data}{a data frame containing the necessary columns (participant_id, titer)}

\item{part_col}{a vector identifying participant id}

\item{val_col}{a vector of target values. Eg. titers}

\item{group_col}{a vector identifying the sample group, default is NULL}

\item{var_trans}{whether the input titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{output_trans}{whether the output titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{base}{log base used to calculate the log transformed value, default is 2}

\item{adj}{dividing factor before taking log transformation, default is 10}

\item{max_titer_col}{the column name of the max titer column, default is "Max_titer"}
}
\value{
a data frame containing the individual max_titer values
}
\description{
Function to format data and calculate max_titer
}
\examples{
\donttest{
data(Fonville)
max_titer_data(Fonville, part_col = "Subject Number", val_col = "titer",
              group_col = "Sample Year")
}

}
