% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_titerData.R
\name{max_titerCal}
\alias{max_titerCal}
\title{Calculate the max_titer for a given participant}
\usage{
max_titerCal(
  i,
  j,
  data,
  part_col,
  val_col,
  group_col = NULL,
  var_trans,
  output_trans,
  base,
  adj
)
}
\arguments{
\item{i}{a participant ID}

\item{j}{a sample group number}

\item{data}{the input data frame}

\item{part_col}{a vector identifying participant id}

\item{val_col}{a vector of target values. Eg. titers}

\item{group_col}{a vector identifying the sample group}

\item{var_trans}{whether the input titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{base}{a parameter passed to the `max_titer` function}

\item{adj}{a parameter passed to the `max_titer` function}

\item{output.log.trans}{whether the output titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}
}
\value{
the max_titer value for the given participant
}
\description{
Calculate the max_titer for a given participant
}
