% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overallData.R
\name{overallData}
\alias{overallData}
\title{Function to format all required metrics of the data}
\usage{
overallData(
  data,
  part_col,
  weight_col,
  val_col,
  group_col = NULL,
  min.y = NULL,
  weight = c(0, 0.05, 0.1, 0.3, 0.6, 0.75, 0.85, 0.9, 0.95, 0.99),
  var_trans = 1,
  max_titer_output_trans = 1,
  gmt_output_trans = 1,
  weight_trans = 1,
  base = 2,
  adj = 10,
  mode = NULL,
  adjust = NULL,
  required_metrics = c("ATY", "AUC", "gini_coefficient", "GMT", "Kurtosis", "Max_titer",
    "Prop2", "Prot_prop", "Skewness", "Width2"),
  col_names = NULL
)
}
\arguments{
\item{data}{a data frame}

\item{part_col}{a vector identifying participant id}

\item{weight_col}{a vector of the distance used to sorting titers. Eg. year of isolation or genetic distance}

\item{val_col}{a vector of target values. Eg. titers}

\item{group_col}{a vector identifying the sample group, default is NULL}

\item{min.y}{the smallest titer value(s), default is the minimum value in the titer vector(calculated later, temporarily being NULL in the formula)}

\item{weight}{a vector of weight values according to titer values, default is c(0.00,0.05,0.10,0.30,0.60,0.75,0.85,0.90,0.95,0.99)}

\item{var_trans}{whether the input titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{max_titer_output_trans}{whether the output max titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{gmt_output_trans}{whether the output gmt value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{weight_trans}{whether the input weight value is transformed to percentages, default is TRUE
TRUE = percentage transformed, FALSE = not percentage transformed}

\item{base}{log base used to calculate the log transformed value, default is 2}

\item{adj}{dividing factor before taking log transformation, default is 10}

\item{mode}{how to calculate values if weight_col is repeated, default is NULL}

\item{adjust}{how to adjust titers, default is the minimum of titers}

\item{required_metrics}{a list of metrics that required by the user. Note that when inputting proportion and width, it should be in the format of "Prop"/"Width" followed by a log_transformed threshold, eg. "Prop2"/"Width2". The other metrics should be exactly "ATY","AUC","gini_coefficient","GMT","Kurtosis","Max_titer","Prot_prop", or "Skewness", i.e. inputting "aty","gini", etc. is considered invalid.}

\item{col_names}{the column names of the output metrics, default is required_metrics}
}
\value{
a data frame containing all metrics
}
\description{
Function to format all required metrics of the data
}
