% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propData.R
\name{propCal}
\alias{propCal}
\title{Calculate the weighted prop for a given participant}
\usage{
propCal(
  i,
  j,
  data,
  part_col,
  val_col,
  group_col,
  var_trans,
  threshold_trans,
  base,
  adj,
  threshold
)
}
\arguments{
\item{i}{a participant ID}

\item{j}{a sample group number}

\item{data}{the input data frame}

\item{part_col}{a vector identifying participant id}

\item{val_col}{a vector of target values. Eg. titers}

\item{group_col}{a vector identifying the sample group, default is NULL}

\item{var_trans}{whether the input titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{threshold_trans}{whether the threshold value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{base}{log base used to calculate the log transformed value, default is 2}

\item{adj}{dividing factor before taking log transformation, default is 10}

\item{threshold}{the titers above which are counted, defalut is 2}
}
\value{
the weighted prop for the given participant
}
\description{
Calculate the weighted prop for a given participant
}
