% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot_prop.R
\name{prot_prop}
\alias{prot_prop}
\title{Calculate the proportion of titers weighted by protection levels}
\usage{
prot_prop(
  titer,
  min_titer = NULL,
  weight = c(0, 0.05, 0.1, 0.3, 0.6, 0.75, 0.85, 0.9, 0.95, 0.99),
  input.log.trans = TRUE,
  input.pct.trans = TRUE,
  base = 2,
  adj = 10
)
}
\arguments{
\item{titer}{a vector of titer values}

\item{min_titer}{the smallest titer value(s), default is the minimum value in the titer vector}

\item{weight}{a vector of weight values according to titer values, default is c(0.00,0.05,0.10,0.30,0.60,0.75,0.85,0.90,0.95,0.99).}

\item{input.log.trans}{whether the input titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{input.pct.trans}{whether the input weight value is transformed to percentages, default is TRUE
TRUE = percentage transformed, FALSE = not percentage transformed}

\item{base}{log base used to calculate the log transformed value, default is 2}

\item{adj}{dividing factor before taking log transformation, default is 10}
}
\value{
prot_prop
}
\description{
Calculate the proportion of titers weighted by protection levels
}
