% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot_propData.R
\name{prot_propCal}
\alias{prot_propCal}
\title{Calculate the weighted prot_prop for a given participant}
\usage{
prot_propCal(
  i,
  j,
  data,
  part_col,
  val_col,
  group_col,
  min.y,
  weight,
  var_trans,
  weight_trans,
  base,
  adj
)
}
\arguments{
\item{i}{a participant ID}

\item{j}{a sample group number}

\item{data}{the input data frame}

\item{part_col}{a vector identifying participant id}

\item{val_col}{a vector of target values. Eg. titers}

\item{group_col}{a vector identifying the sample group, default is NULL}

\item{min.y}{the smallest titer value(s), default is the minimum value in the titer vector}

\item{weight}{a vector of weight values according to titer values, default is c(0.00,0.05,0.10,0.30,0.60,0.75,0.85,0.90,0.95,0.99)}

\item{var_trans}{whether the input titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{weight_trans}{whether the input weight value is transformed to percentages, default is TRUE
TRUE = percentage transformed, FALSE = not percentage transformed}

\item{base}{log base used to calculate the log transformed value, default is 2}

\item{adj}{dividing factor before taking log transformation, default is 10}
}
\value{
the weighted prot_prop for the given participant
}
\description{
Calculate the weighted prot_prop for a given participant
}
