% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot_propData.R
\name{prot_propData}
\alias{prot_propData}
\title{Function to format data and calculate the weighted prot_prop}
\usage{
prot_propData(
  data,
  part_col,
  val_col,
  group_col = NULL,
  min.y = NULL,
  weight = c(0, 0.05, 0.1, 0.3, 0.6, 0.75, 0.85, 0.9, 0.95, 0.99),
  var_trans = 1,
  weight_trans = 1,
  base = 2,
  adj = 10,
  prot_prop_col = "Prot_prop"
)
}
\arguments{
\item{data}{a data frame containing the necessary columns (participant_id titer)}

\item{part_col}{a vector identifying participant id}

\item{val_col}{a vector of target values. Eg. titers}

\item{group_col}{a vector identifying the sample group, default is NULL}

\item{min.y}{the smallest titer value(s), default is the minimum value in the titer vector(calculated later, temporarily being NULL in the formula)}

\item{weight}{a parameter passed to the \code{prot_prop} function}

\item{var_trans}{whether the input titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{weight_trans}{whether the input weight value is transformed to percentages, default is TRUE
TRUE = percentage transformed, FALSE = not percentage transformed}

\item{base}{log base used to calculate the log transformed value, default is 2}

\item{adj}{dividing factor before taking log transformation, default is 10}

\item{prot_prop_col}{the column name of the protected proportion column, default is "Prot_prop"}
}
\value{
a data frame containing the individual weighted prot_prop
}
\description{
Function to format data and calculate the weighted prot_prop
}
