% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeroMetrics-package.R
\docType{package}
\name{SeroMetrics-package}
\alias{SeroMetrics}
\alias{SeroMetrics-package}
\title{SeroMetrics: Summary Metrics for Antibody Landscapes}
\description{
Quantitatively summarizes antibody landscapes using metrics that capture the magnitude, extent, skewness, and tailedness of antibody response profiles. Functions are provided to calculate these metrics from titer data and publicly available datasets are included for demonstration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bingyiyang/SeroMetrics}
  \item Report bugs at \url{https://github.com/bingyiyang/SeroMetrics/issues}
}

}
\author{
\strong{Maintainer}: Bingyi Yang \email{byyang@connect.hku.hk}

Authors:
\itemize{
  \item Shixian Xu
}

}
