% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness_data.R
\name{skewness_data}
\alias{skewness_data}
\title{Function to format data and calculate skewness}
\usage{
skewness_data(
  data,
  part_col,
  weight_col,
  val_col,
  group_col = NULL,
  var_trans = TRUE,
  base = 2,
  adj = 10,
  mode = NULL,
  adjust = NULL,
  skew_col = "Skewness"
)
}
\arguments{
\item{data}{a data frame containing the necessary columns (participant_id, titer)}

\item{part_col}{a vector identifying participant id}

\item{weight_col}{a vector of the distance used to sorting titers. Eg. year of isolation or genetic distance}

\item{val_col}{a vector of target values. Eg. titers}

\item{group_col}{a vector identifying the sample group, default is NULL}

\item{var_trans}{whether the input titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{base}{log base used to calculate the log transformed value, default is 2}

\item{adj}{dividing factor before taking log transformation, default is 10}

\item{mode}{how to calculate values if weight_col is repeated, default is NULL}

\item{adjust}{how to adjust titers, default is the minimum of titers}

\item{skew_col}{the column name of the skewness column, default is "Skewness"}
}
\value{
a data frame containing the individual skewness values
}
\description{
Function to format data and calculate skewness
}
\examples{
\donttest{
data(Fonville)
skewness_data(Fonville, part_col = "Subject Number", weight_col = "isolation_year",
             val_col = "titer", group_col = "Sample Year")
}

}
