% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniq.R
\name{uniq}
\alias{uniq}
\title{Select the samples according to the requirements}
\usage{
uniq(data, columns = NULL)
}
\arguments{
\item{data}{data frame}

\item{columns}{columns to select}
}
\value{
a data frame of unique rows for the selected columns
}
\description{
Select the samples according to the requirements
}
\examples{
df <- data.frame(id = c(1, 1, 2), year = c(2000, 2000, 2001),
                 titer = c(5, 5, 6))
uniq(df)
uniq(df, columns = c("id", "year"))

}
