% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/width.R
\name{width}
\alias{width}
\title{Calculate the width of the titers}
\usage{
width(
  titer,
  distance.weight,
  threshold = 2,
  threshold_trans = TRUE,
  input.log.trans = TRUE,
  base = 2,
  adj = 10,
  adjust = NULL
)
}
\arguments{
\item{titer}{titer value}

\item{distance.weight}{a vector the distance used to sorting titers. Eg. year of isolation or amino acid distance}

\item{threshold}{the titers above which are counted, default is 2}

\item{threshold_trans}{whether the threshold value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{input.log.trans}{whether the input titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{base}{log base used to calculate the log transformed value, default is 2}

\item{adj}{dividing factor before taking log transformation, default is 10
#' @param adjust how to adjust titers, default is the minimum of titers}
}
\value{
width
}
\description{
Calculate the width of the titers
}
