% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widthData.R
\name{widthData}
\alias{widthData}
\title{Function to format data and calculate the weighted width}
\usage{
widthData(
  data,
  part_col,
  weight_col,
  val_col,
  group_col = NULL,
  var_trans = 1,
  threshold_trans = 1,
  base = 2,
  adj = 10,
  threshold = 2,
  mode = NULL,
  adjust = NULL
)
}
\arguments{
\item{data}{a data frame containing the necessary columns (participant_id, isolation_year, titer)}

\item{part_col}{a vector identifying participant id}

\item{weight_col}{a vector of the distance used to sorting titers. Eg. year of isolation or genetic distance}

\item{val_col}{a vector of target values. Eg. titers}

\item{group_col}{a vector identifying the sample group, default is NULL}

\item{var_trans}{whether the input titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{threshold_trans}{whether the threshold value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{base}{log base used to calculate the log transformed value, default is 2}

\item{adj}{dividing factor before taking log transformation, default is 10}

\item{threshold}{the titers above which are counted, default is 2}

\item{mode}{how to calculate values if weight_col is repeated, default is NULL}

\item{adjust}{how to adjust titers, default is the minimum of titers}
}
\value{
a data frame containing the individual weighted width
}
\description{
Function to format data and calculate the weighted width
}
