% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.R
\name{prop}
\alias{prop}
\title{Calculate proportion of titers over a threshold}
\usage{
prop(
  titer,
  threshold = 2,
  threshold.log.trans = TRUE,
  input.log.trans = TRUE,
  base = 2,
  adj = 10
)
}
\arguments{
\item{titer}{titer value#'}

\item{threshold}{the titers above which are counted, defalut is 2}

\item{threshold.log.trans}{whether the threshold value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{input.log.trans}{whether the input titer value is log transformed, default is TRUE
TRUE = log transformed, FALSE = not log transformed}

\item{base}{log base used to calculate the log transformed value, default is 2}

\item{adj}{dividing factor before taking log transformation, default is 10}
}
\value{
PROP
}
\description{
Calculate proportion of titers over a threshold
}
