% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniq.R
\name{uniq}
\alias{uniq}
\title{Select the samples according to the requirements}
\usage{
uniq(data, columns = NULL)
}
\arguments{
\item{data}{data frame}

\item{columns}{columns to select}
}
\value{
uniq
}
\description{
Select the samples according to the requirements
}
